Rename : 로딩 컴포넌트 이동
